+++ src/stores/index.js
@@ -1,6 +1,2 @@
 // From the stores folder
-export AppStore from './AppStore.js';
 export StyleStore from './StyleStore.js';
-
-// Stores from other folders
-export ViewsStore from 'components/views/Views.store';